databaese config

+++ clients/src/main/java/com/edenlifemock/clients/food/MealObject.java
@@ -7,6 +7,6 @@ public record MealObject(
          String mealName,
          String type,
          String desc,
-         BigDecimal price
+         Double price
 ) {
 }
