+++ clients/src/main/java/com/edenlifemock/clients/food/UpdateMealRequest.java
@@ -1,10 +1,8 @@
 package com.edenlifemock.clients.food;
 
-import java.math.BigDecimal;
-
 public record UpdateMealRequest(Long mealId,
-         String mealName,
-         String type,
-         String desc,
-         BigDecimal price) {
+                                String mealName,
+                                String type,
+                                String desc,
+                                Double price) {
 }
